absence req changes

+++ AddOptimization.Contracts/Dto/LeaveBalanceDto.cs
@@ -10,8 +10,8 @@ namespace AddOptimization.Contracts.Dto
     {
         public int EmployeeId { get; set; }
         public int TotalAllocatedHoliday {  get; set; }
-        public  int LeaveTaken  { get; set; }
-        public int leavesLeft { get; set; }
+        public  Decimal LeaveTaken  { get; set; }
+        public Decimal leavesLeft { get; set; }
 
     }
 }
